+++ docs/pages/x/api/date-pickers/date-range-picker-day.json
@@ -99,8 +99,7 @@
         "name": "union",
         "description": "func<br>&#124;&nbsp;{ current?: { pulsate: func, start: func, stop: func } }"
       }
-    },
-    "type": { "type": { "name": "string" }, "default": "'button'" }
+    }
   },
   "name": "DateRangePickerDay",
   "imports": [
