+++ docs/pages/x/api/date-pickers/picker-day.json
@@ -92,8 +92,7 @@
         "name": "union",
         "description": "func<br>&#124;&nbsp;{ current?: { pulsate: func, start: func, stop: func } }"
       }
-    },
-    "type": { "type": { "name": "string" }, "default": "'button'" }
+    }
   },
   "name": "PickerDay",
   "imports": [
