+++ docs/translations/api-docs/date-pickers/date-range-picker-day/date-range-picker-day.json
@@ -117,9 +117,6 @@
     "TouchRippleProps": { "description": "Props applied to the <code>TouchRipple</code> element." },
     "touchRippleRef": {
       "description": "A ref that points to the <code>TouchRipple</code> element."
-    },
-    "type": {
-      "description": "The HTML <a href=\"https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements/button#type\"><code>type</code></a> attribute applied to <code>button</code> and <code>a</code> elements. Ignored when rendering non-native buttons."
     }
   },
   "classDescriptions": {
