+++ packages/x-date-pickers-pro/src/DateRangePickerDay/DateRangePickerDay.tsx
@@ -768,13 +768,6 @@ DateRangePickerDayRaw.propTypes = {
       }),
     }),
   ]),
-  /**
-   * The HTML [`type`](https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements/button#type)
-   * attribute applied to `button` and `a` elements.
-   * Ignored when rendering non-native buttons.
-   * @default 'button'
-   */
-  type: PropTypes.string,
 } as any;
 
 /**
