+++ packages/x-date-pickers/src/PickerDay/PickerDay.tsx
@@ -482,13 +482,6 @@ PickerDayRaw.propTypes = {
       }),
     }),
   ]),
-  /**
-   * The HTML [`type`](https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements/button#type)
-   * attribute applied to `button` and `a` elements.
-   * Ignored when rendering non-native buttons.
-   * @default 'button'
-   */
-  type: PropTypes.string,
 } as any;
 
 /**
