+++ packages/x-scheduler-premium/src/internals/components/recurring-scope-dialog/RecurringScopeDialog.tsx
@@ -14,16 +14,16 @@ import { RecurringEventUpdateScope } from '@mui/x-scheduler-internals/models';
 import { schedulerOtherSelectors } from '@mui/x-scheduler-internals/scheduler-selectors';
 import { useSchedulerStoreContext } from '@mui/x-scheduler-internals/use-scheduler-store-context';
 import { useEventDialogStyledContext } from '@mui/x-scheduler/internals';
-import { RecurringScopeDialogProps } from './ScopeDialog.types';
+import { RecurringScopeDialogProps } from './RecurringScopeDialog.types';
 
 export const RecurringScopeDialog = React.forwardRef<HTMLDivElement, RecurringScopeDialogProps>(
-  function ScopeDialog(props, ref) {
+  function RecurringScopeDialog(props, ref) {
     // Context hooks
     const { schedulerId, localeText } = useEventDialogStyledContext();
     const store = useSchedulerStoreContext();
 
     // Selector hooks
-    const open = useStore(store, schedulerOtherSelectors.isScopeDialogOpen);
+    const open = useStore(store, schedulerOtherSelectors.isRecurringScopeDialogOpen);
 
     // Feature hooks
     const handleOpenChange = React.useCallback(
