+++ packages/x-scheduler-premium/src/internals/eventDialogOptionalRenderers.ts
@@ -1,5 +1,5 @@
 import { RecurrenceTab } from './components/event-dialog/RecurrenceTab';
-import { RecurringScopeDialog } from './components/scope-dialog/ScopeDialog';
+import { RecurringScopeDialog } from './components/recurring-scope-dialog/RecurringScopeDialog';
 
 export const PREMIUM_EVENT_DIALOG_OPTIONAL_RENDERERS = {
   recurrenceTab: RecurrenceTab,
