+++ packages/x-scheduler-premium/src/internals/index.ts
@@ -1,3 +1,3 @@
 export { RecurrenceTab } from './components/event-dialog/RecurrenceTab';
-export { RecurringScopeDialog } from './components/scope-dialog/ScopeDialog';
+export { RecurringScopeDialog } from './components/recurring-scope-dialog/RecurringScopeDialog';
 export { PREMIUM_EVENT_DIALOG_OPTIONAL_RENDERERS } from './eventDialogOptionalRenderers';
