+++ packages/x-scheduler/src/locales/esES.ts
@@ -57,7 +57,7 @@ const esESDialog: Partial<EventDialogLocaleText> = {
   startDateLabel: 'Fecha de inicio',
   startTimeLabel: 'Hora de inicio',
 
-  // ScopeDialog
+  // RecurringScopeDialog
   all: 'Todos los eventos',
   cancel: 'Cancelar',
   confirm: 'Confirmar',
