+++ packages/x-scheduler/src/locales/plPL.ts
@@ -56,7 +56,7 @@ const plPLDialog: Partial<EventDialogLocaleText> = {
   // startDateAfterEndDateError: 'Start date/time must be before end date/time.',
   // startDateLabel: 'Start date',
   // startTimeLabel: 'Start time',
-  // ScopeDialog
+  // RecurringScopeDialog
   // all: 'All events',
   // cancel: 'Cancel',
   // confirm: 'Confirm',
