+++ packages/x-scheduler/src/locales/roRO.ts
@@ -57,7 +57,7 @@ const roRODialog: Partial<EventDialogLocaleText> = {
   startDateLabel: 'Data de început',
   startTimeLabel: 'Ora de început',
 
-  // ScopeDialog
+  // RecurringScopeDialog
   all: 'Toate evenimentele',
   cancel: 'Anulează',
   confirm: 'Confirmă',
